requirements.txt
streamlit==1.38.0
requests==2.32.3
Pillow==10.4.0
sounddevice
scipy
python-dotenv==1.0.1